backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "assistant-stream[langgraph]>=0.0.30",
    "bson>=0.5.10",
    "fastapi[standard]>=0.119.1",
    "fastmcp>=2.12.5",
    "langchain>=1.0.2",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "mcp>=1.16.0",
    "openai>=2.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "typer>=0.20.0",
    "xlrd>=2.0.2",
]
